Add files via upload
This is an interactive dashboard of a marketing organization. Regards to the sales data, it is a walk-into retail store that provides insight into Sylip Corporation, a multifaceted retail firm that sells groceries, clothes, household goods, and extra. The project concentrates on crucial key performance indicators (KPIs) using sales data to offer practical insights by carefully cleaning, manipulating, analysing, and visualizing the data. The project provides a thorough insight of the company's success by disclosing region-specific earnings, annual revenue, and annual unit sold. Making well-informed decisions will help Sylip Corporation increase optimization and promote overall success. The project produced a thorough study of sales data, insightful reports for the business, and suggestions for further business optimization potential.


From the period under review, the total revenue made between 2010-2017 from all regions, were £137.35M, total cost was £93.18M and total profit were £44.17M approximately. The rate of annual total revenue was presented in this chart according to the KPIs. It was revealed that 13.97% was achieved in 2010 which further declined in 2011 to 8.4%. 2012 recorded a significant increase in total revenue with the rate of 23.22%. It was also reported that SYLIP Corporation made a decline in its revenue in 2013 down to 2016 while there was a slight increase in the revenue in 2017 with the value of 9.74%. 

Furthermore, the total profit achieved was £44,168,198.40. The region with the highest profit made was Sub-Saharan Africa with £12.18M, Europe with £ 11.08M, Asia with £6.11M respectively. It was indicated that Sub-Saharan has the highest profit while North America was reported to have lowest profit. The four top Countries with the highest revenue were Honduras, and Myanmar, Dijbouti, and Turkmenistan. The countries with the lowest revenue made were Kyrgyzstan, Slovakia, Kuwait, and Comoros. 
 
Recommendation was made which one of them were Sylip Corporation should evaluate its retail business to determine which items are most suited for sales in the countries that scored the lowest performance levels. These entails determining if the products that sold in those countries are necessities for the average consumer. Using strategic approaches to investigate and implore representatives to conduct thorough surveys can increase business optimization in the geographical countries that have low sales performance. Furthermore, Sylip Corporation should focus more on the countries that generate more revenue.