Create README.md
This Dashboard uncovers students who signed up for online career courses through Google Ads. For those who were interested, there were 11 Program offered. Business analysis, project management, data analysis, software development, digital marketing, manual and automated testing, and scrum mastering are among the topics covered in these courses. The goal was to ascertain which channels drove the most website traffic and which courses, out of all those advertised, had the highest enrolment rates. The students learned about the classes through a variety of media, including Twitter, Instagram, TikTok, Facebook YouTube, and friend referrals. With 46.3% of all enrolments, Business Analysis leads the league of programs with the highest sign-up rate, according to the data. With a combined share of 94%, the top 4 courses are Business Analysis, Project Management, Data Analysis, and Manual Testing. With a combined share of barely 1%, Scrum Master, Graphic Design, and Risk Management are the three least effective courses or programs.
When it came to the channel evaluation of "Where did the student hear about us," "Friend" yielded the highest positive return, with 48% of enrollees choosing a course because of a friend's recommendation. Additionally, the cumulative share of the top 4 channels, which are YouTube, Google Search, Instagram Ads, and Friends, is 66%. collectively accounted for 85.7% of all enrolments, the majority of which took place in 2019 and 2021.
I recommend the business managers to examine ways to increase sales from the most popular programs and to assess other programs to find potential ways to increase subscriptions from them. Program coordinators need to make sure that everyone receives high-quality service because this leads to happy customers and "word-of-mouth" referrals from friends, a channel that generates more revenue than spending money on YouTube advertisements.